Move useTextData to hooks
