+++ packages/lib-classifier/src/helpers/index.js
@@ -9,7 +9,6 @@ export { default as shownMarks } from './shownMarks'
 export { default as subjectsSeenThisSession } from './subjectsSeenThisSession'
 export { default as subjectViewers } from './subjectViewers'
 export { default as useJSONData } from './useJSONData'
-export { default as useTextData } from './useTextData'
 export { default as validateMimeType } from './validateMimeType'
 export { default as validateSubjectLocations } from './validateSubjectLocations'
 export { default as withFeatureFlag } from './withFeatureFlag'
