--- packages/lib-classifier/src/helpers/useTextData/index.js
@@ -1,1 +1,0 @@
-export { default } from './useTextData'
