+++ packages/lib-classifier/src/hooks/index.js
@@ -6,5 +6,6 @@ export { default as usePanoptesUser } from './usePanoptesUser'
 export { default as useProjectPreferences } from './useProjectPreferences'
 export { default as useProjectRoles } from './useProjectRoles'
 export { default as useStores } from './useStores'
+export { default as useTextData } from './useTextData'
 export { default as useTranscriptionReductions } from './useTranscriptionReductions'
 export { default as useWorkflowSnapshot } from './useWorkflowSnapshot'
